+++ packages/jaspr/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased patch
 
+- Fixed flutter embedding issue with version 3.29.0.
 - Added `onClick` override to `a` html tag. Fixing issues with default action handling.
 - Added `--extra-js-compiler-option` and `--extra-wasm-compiler-option` to `jaspr build` command.